backend/requirements.txt
fastapi
uvicorn
pydantic-settings
python-dotenv
python-multipart
langchain
langchain-core
langchain-groq
langchain-community
langchain-google-genai
langsmith
tavily-python
langgraph
pymupdf
langchain-huggingface 
sentence-transformers
chromadb
numpy
google-genai
langchain-tavily